requirements.txt
Django==2.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-guardian==1.4.9
djangorestframework==3.7.3
Markdown==2.6.10
pytz==2017.3
six==1.11.0
